functions to conftests
